# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
        "/api/registration": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Register",
                "description": "Login by phone number, password",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password",
                                    "first_name",
                                    "last_name",
                                    "country_code",
                                    "gender",
                                    "birthdate",
                                    "avatar",
                                    "phone_number"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    },
                                    "avatar": {
                                        "type": "string",
                                        "format": "base64",
                                        "example": "data:image/jpeg;base64, yourSuperLongStringBinary"
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "example": "ahmed"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "adel"
                                    },
                                    "country_code": {
                                        "type": "string",
                                        "example": "EG"
                                    },
                                    "birthdate": {
                                        "type": "string",
                                        "example": "+201095222457"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "format": "E.164",
                                        "example": "+201095222457"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "example": "EG"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/signIn": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "LinkStatus",
                "description": "link status with user",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password",
                                    "phone_number"
                                ],
                                "properties": {
                                    "phone_number": {
                                        "type": "string",
                                        "example": "+201095111472"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong phone number address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/createStatus": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "LinkStatus",
                "description": "link status with user",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status",
                                    "phone_number"
                                ],
                                "properties": {
                                    "phone_number": {
                                        "type": "string",
                                        "example": "+201096444782"
                                    },
                                    "status": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "test status"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong phone number address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    }
}